+++ 2_transform/commons/logging/__init__.py
@@ -1,3 +1,3 @@
-from .logger import get_logger
+from .logger import configure_flow_logging, get_logger
 
-__all__ = ["get_logger"]
+__all__ = ["get_logger", "configure_flow_logging"]